Verify initial amount in genesis

+++ core/src/main/java/com/radixdlt/rev2/REv2StateReader.java
@@ -66,7 +66,8 @@ package com.radixdlt.rev2;
 
 import com.radixdlt.address.ComponentAddress;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public interface REv2StateReader {
-  BigDecimal getComponentResources(ComponentAddress componentAddress);
+  BigInteger getComponentResources(ComponentAddress componentAddress);
 }
